% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr_mlsclust}
\alias{fdr_mlsclust}
\title{Compute the false discovery rate (FDR) of \emph{mlscluster}.}
\usage{
fdr_mlsclust(
  path_stats,
  cog_md_muts_syn_ranges,
  codon_3rd_pos,
  codon_1st_2nd_pos,
  out_folder
)
}
\arguments{
\item{path_stats}{Path with the output files generated by \code{\link[=run_diff_thresholds]{run_diff_thresholds()}}. Identical to the supplied
\code{output_dir} in the \code{\link[=stats_multiple_thresholds]{stats_multiple_thresholds()}} function (default: NULL)}

\item{cog_md_muts_syn_ranges}{First list element returned by \code{\link[=compute_muts_match_codons]{compute_muts_match_codons()}}}

\item{codon_3rd_pos}{Second list element returned by \code{\link[=compute_muts_match_codons]{compute_muts_match_codons()}}}

\item{codon_1st_2nd_pos}{Third list element returned by \code{\link[=compute_muts_match_codons]{compute_muts_match_codons()}}}

\item{out_folder}{Simular to parameter with same name from \code{\link[=stats_multiple_thresholds]{stats_multiple_thresholds()}}. Outputs will be generated
inside \verb{stat_results/\{out_folder\}/fdr_codons/}}
}
\value{
Arranged plots of FDR and epsilon overall and across different genomic regions
}
\description{
A key assumption is that synonymous mutations do not provide a fitness advantage
and would then represent an erroneous TFP call. Since mutations occur much more frequently
in 3rd codon positions, FDR is calculated only considering the proportion of erroneous calls
across polymorphic 3rd codon mutated sites. A more conservative approach (epsilon) considers
both the FDR and the erroneous proportion of calls on sites at first and second codon positions.
}
