% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsclust.R
\name{mlsclust}
\alias{mlsclust}
\title{Title}
\usage{
mlsclust(
  tre,
  amd,
  min_descendants = 10,
  max_descendants = 20000,
  min_cluster_age_yrs = 1/12,
  min_date = NULL,
  max_date = NULL,
  branch_length_unit = "years",
  defining_mut_threshold = 0.75,
  root_on_tip = "Wuhan/WH04/2020",
  root_on_tip_sample_time = 2019.995,
  detailed_output = FALSE,
  ncpu = 1
)
}
\arguments{
\item{tre}{}

\item{amd}{}

\item{min_descendants}{}

\item{max_descendants}{}

\item{min_cluster_age_yrs}{}

\item{min_date}{}

\item{max_date}{}

\item{branch_length_unit}{}

\item{defining_mut_threshold}{}

\item{root_on_tip}{}

\item{root_on_tip_sample_time}{}

\item{detailed_output}{}

\item{ncpu}{}
}
\value{

}
\description{
Title
}
