% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_thresholds.R
\name{plot_tbc_stats}
\alias{plot_tbc_stats}
\title{Plot boxplot of the clustering statistics}
\usage{
plot_tbc_stats(
  path_stats,
  out_folder,
  ymin_vec = c(0, 0, -50),
  ymax_vec = c(10, 10, 100)
)
}
\arguments{
\item{path_stats}{Path with the output files generated by \code{\link[=run_diff_thresholds]{run_diff_thresholds()}}. Identical to the supplied
\code{output_dir} in the latter function (default: NULL)}

\item{out_folder}{Output folder which will be appended to \verb{stat_results/} in the working directory to
generate results for multiple thresholds. E.g. \code{out_folder=="period2"} then results will go to
\code{stat_results/period2/threshold_quantile_0.25}, and the other 9 thresholds (0.5, 0.75, 1, 2, 3,
4, 5, 10, 25). Default: NULL}

\item{ymin_vec}{Manual adjustment of minimum values for (i) ratio sizes, (ii) ratio persistence time,
and (iii) logistic growth rate}

\item{ymax_vec}{Same as above but adjusting for maximum values}
}
\value{
An arranged figure with 3 rows (one for each statistic) containing boxplots of stat values for each threshold
}
\description{
Plot boxplot of the clustering stats values across clades detected by at least one of the stats (union) for different thresholds
}
