% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_thresholds.R
\name{stacked_nsites_genomic_region_mult_thresholds}
\alias{stacked_nsites_genomic_region_mult_thresholds}
\title{Plot the number of TFP-homoplasies by genomic region coloured by the first threshold detected}
\usage{
stacked_nsites_genomic_region_mult_thresholds(path)
}
\arguments{
\item{path}{Directory path of the form \verb{stat_results/\{out_folder\}/genomewide_plot_non-syn/} obtained
after running \code{\link[=stats_multiple_thresholds]{stats_multiple_thresholds()}}}
}
\value{
Plot TFP-homoplasy sites by genomic region coloured by the first threshold detected.
This is also generated at \verb{\{path\}/stacked_nsites_threshold/}
}
\description{
Plot the number of TFP-homoplasies by genomic region coloured by the first threshold detected
}
