% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_thresholds.R
\name{stats_multiple_thresholds}
\alias{stats_multiple_thresholds}
\title{Title}
\usage{
stats_multiple_thresholds(
  path_stats,
  rm_freq_outliers = TRUE,
  out_folder,
  out_poisson_file
)
}
\arguments{
\item{path_stats}{}

\item{rm_freq_outliers}{}

\item{out_folder}{}

\item{out_poisson_file}{}
}
\value{

}
\description{
Title
}
