% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_threshold.R
\name{stats_selected_threshold}
\alias{stats_selected_threshold}
\title{Return TFP-homoplasy plots for the selected threshold after FDR inspection}
\usage{
stats_selected_threshold(
  path_stats,
  rm_freq_outliers = TRUE,
  thr_index = 5,
  out_folder
)
}
\arguments{
\item{path_stats}{Path with the output files generated by \code{\link[=run_diff_thresholds]{run_diff_thresholds()}}. Identical to the supplied
\code{output_dir} in the latter function (default: NULL)}

\item{rm_freq_outliers}{Whether to remove >99\% quantile threshold of Freq_homopl (usually related to
sequencing artifacts). Default: TRUE}

\item{thr_index}{Index of the cluster/quantile threshold selected for targeted analysis.
The default thresholds are: 0.25, 0.5, 0.75, 1, 2, 3, 4, 5, 10, and 25\%.
Therefore, the \code{thr_index} for threshold 2\% is 5 (default).}

\item{out_folder}{Output folder which will be appended to \verb{stat_results/} in the working directory
to generate results for the selected threshold. Should be different from the folder specified at
\code{\link[=stats_multiple_thresholds]{stats_multiple_thresholds()}} to avoid overriding results.}
}
\value{
A list with 4 elements:
\itemize{
\item \code{sbf_s} generates a stacked bar plot of unique synonymous
homoplasies under multilevel selection (TFPs) by major lineage and genomic region
\item \code{sbf_ns} returns the same as \code{sbf_s} but for non-synonymous TFP-homoplasies
\item \code{gp_ns} return plots of non-synonymous TFP-homoplasy frequencies along each genomic
region coloured by major lineage
\item \code{bubble_non_syn_clust1} returns (i) a plot comparing sites identified by mlscluster
(multilevel selection) and a dN/dS HyPhy approach (positive selection), and
(ii) a bubble plot of TFP-homoplasy frequencies stratified by major lineage
}
}
\description{
Return TFP-homoplasy plots for the selected threshold after FDR inspection
}
