% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{compute_muts_match_codons}
\alias{compute_muts_match_codons}
\title{Title}
\usage{
compute_muts_match_codons(cog_md_muts, nseqs)
}
\arguments{
\item{cog_md_muts}{}

\item{nseqs}{}
}
\value{

}
\description{
Title
}
