% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr_mlsclust}
\alias{fdr_mlsclust}
\title{Title}
\usage{
fdr_mlsclust(
  path_stats,
  cog_md_muts_syn_ranges,
  codon_3rd_pos,
  codon_1st_2nd_pos,
  out_folder
)
}
\arguments{
\item{path_stats}{}

\item{cog_md_muts_syn_ranges}{}

\item{codon_3rd_pos}{}

\item{codon_1st_2nd_pos}{}

\item{out_folder}{}
}
\value{

}
\description{
Title
}
