% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_thresholds.R
\name{stats_multiple_thresholds}
\alias{stats_multiple_thresholds}
\title{Function to facilitate comparison between periods and across thresholds}
\usage{
stats_multiple_thresholds(
  path_stats,
  pal_lineages = c(Alpha_B.1.1.7 = "#fd7f6f", `Delta_AY.4.*` = "#7eb0d5", Delta_other =
    "#b2e061", EU1_B.1.177 = "#bd7ebe", `Omicron_BA.1.*` = "#ffb55a", `Omicron_BA.2.*` =
    "#ffcc00", Other = "#fdcce5"),
  out_folder,
  out_poisson_file
)
}
\arguments{
\item{path_stats}{Path with the output files generated by \code{\link[=run_diff_thresholds]{run_diff_thresholds()}}. Identical to the supplied
\code{output_dir} in the latter function (default: NULL)}

\item{pal_lineages}{Customisable palette to use for major_lineage plots. Specified as c("lineage_label"="#color", ...)}

\item{out_folder}{Output folder which will be appended to \verb{stat_results/} in the working directory to
generate results for multiple thresholds. E.g. \code{out_folder=="period2"} then results will go to
\code{stat_results/period2/threshold_quantile_0.25}, and the other 9 thresholds (0.5, 0.75, 1, 2, 3,
4, 5, 10, 25). Default: NULL}

\item{out_poisson_file}{Output file for poisson regressions across thresholds having the Freq of homoplasy
as response and major lineage, genomic region and whether independently found under positive selection
as predictors. Will be appended to \verb{stat_results/\{out_folder\}/}. E.g. assuming again \code{out_folder=="period2"}
and now \code{out_poisson_file=="poisson_p2.txt"} then output will be added to \code{stat_results/period2/poisson_p2.txt}}
}
\value{
A list with 3 elements:
\itemize{
\item \code{bubble_non_syn_clust1} gives TFP-homoplasy frequency bubble plots across thresholds
for each major lineage and intersection with positive selection
\item \code{s9} returns as the first nested list the positioning and frequency of non-synonymous TFP-homoplasies for each genomic
region coloured by the first detected threshold. The second nested list gives the same as above but coloured by major_lineage
\item \code{s13} returns the same as \code{s9} but for synonymous instead of non-synonymous TFP-homoplasies
}
}
\description{
Generate multiple visualizations and perform statistical test to detect major lineages and
genomic regions enriched for TFP-homoplasies
}
