% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_threshold.R
\name{stats_selected_threshold}
\alias{stats_selected_threshold}
\title{Title}
\usage{
stats_selected_threshold(
  path_stats,
  rm_freq_outliers = TRUE,
  thr_index,
  out_folder
)
}
\arguments{
\item{path_stats}{}

\item{rm_freq_outliers}{}

\item{thr_index}{}

\item{out_folder}{}
}
\value{

}
\description{
Title
}
